requirements.txt
bleach==6.0.0
Django>=3.2,<3.3
django-anymail[mailgun]==8.5
django-cors-headers==3.5.0
django-extensions==3.2.3
django-hijack==3.4.1
django-model-utils==4.2.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
factory_boy==3.2.1
getconf==1.11.1
gunicorn==20.1.0
ipython==8.14.0
numpy==1.22.3
pandas==1.4.1
psycopg2==2.9.5
rollbar>=0.16,<0.17
telescoop_backup==0.1.4
wagtail>=5.0,<5.1
wagtailsvg==0.0.37
